init_app/requirements.txt
Django>=3.0,<4.0
djangorestframework >=3.12
Pillow>=4.0.0,<5